requirement.txt
Django==5.0.3
djangorestframework==3.14
django-cors-headers==4.3.0
python-dotenv==1.0.1
psycopg[binary,pool]==3.1.8
